[FIRRTL] Add support for "intrinsics" (#4429)
Add support for intrinsics in FIRRTL. Until intrinsics are supported in the firrtl spec, or we have more complex intrinsics, do direct lowering of intrinsics into their ops.

It is expected that eventually something like an "intmodule" will exist and annotated extern modules will not be used for intrinsics or first they will lower to intmodules internally, then intmodules will be handled uniformly.

As an example, add a sizeof operator which returns the number of bits in the argument type. This let's you query the result of type inference from inside the circuit.

Also adds support for isX as an intrinsic and plusarg sv functions.