[FIRRTL] Change FIRRTL fieldID fns to use uint64_t
Convert all fieldID-related functions which are not already in the
FieldIDTypeInterface to use uint64_t for its type.  Previously, this was
a mix of unsigned for bundles and size_t for vectors (due to reasons
that vectors can sometimes be representing very large memories with more
than 2^32 elements whereas this is rare for bundles).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>