[FIRRTL] Mark DUT AddSeqMemPorts ports dontTouch
Change the AddSeqMemPorts pass to mark ports created on the DUT (or the
top module) with DontTouchAnnotations.  This is done to avoid having
these ports be deleted.  This is currently a problem for output ports
that the pass creates and bores upwards.  The final instance port of the
DUT does not have a user which causes all the ports (except the one on
the memory) to be deletion candidates.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>