[ExportVerilog] Move local name legalization to pre-pass (#4573)
This PR removes ModuleNameManager from ExportVerilog. Previously it was necessary for ExportVerilog to manage local names and perform legalization online since ExportVerilog had to spill temporary wires while emission for syntax legalization. However, we separated all spilling logics into Prepare, it's possible to know all operations which can declare names. 

* GlobalNameResolver was changed to legalize all the names parallelly. 
* Remove ModuleNameManager from ExportVerilog and look up names through `getSymOpName` and `getVerilogValueName` functions. 
* Prepare was modified to return `LogicalResult` so that we can detect unlowered operations more early (this could be merged in a separated commit).

Basically there is no change in output verilog except for a tiny change of port and parameter names. This is caused by the change of the order of name legalization. (we legalized ports first, but this PR changes to legalize module parameters first).