[HW] Fix creating modules with InOut typed ports
When we create the PortInfo object for a module, if the port is a
HW::InOutType, we strip it and set the isInOut field to true.  When we create a
module from a PortInfo object we respect this field while constructing the
function type, but were not respecting the isInOut when creating the block
arguments.

This change starts respecting the `isInOut` field of port info when
constructing the block argument types.  This also adds a verifier that the
function type of the module matches the block argument types.