[CalyxToFSM] Add FSM-flow remove groups pass (#4600)
This commit adds a version of the Calyx remove-groups pass tailored for
the FSM flow. Specifically, it will perform FSM outlining alongside group
rewriting - which interacts with the FSM I/O (group go/done signals).

Technically, we now (finally) have an end-to-end flow for SCF->Calyx->FSM->SV.
The usability of the end-to-end flow is still fairly restricted, mainly
due to:
1. the lack of support for lowering Calyx memory ops
2. The lack of support for lowering `calyx.par` ops in the FSM lowering
   (this shouldn't be too difficult).

A fair amount of state is stored in attributes in between `materialize-calyx-to-fsm`
and `calyx-remove-groups-fsm`. I'm debating with myself whether that's code
smell or just a necessary evil given the approach taken. Nevertheless, I
did this work because of being frustrated with the disconnection of Calyx
from the rest of CIRCT, which renders it unuseable. Since I was the author
of the FSM flow, I figured that i'd might aswell finish the end-to-end
prototype before stepping somewhat away from Calyx-related development.

I'm hoping that given this end-to-end prototype, there's enough existing
infrastructure for others to pick up the torch and flesh out the
missing pieces.

Some additional notes:
* Also makes `calyx.component` contain a graph region to make it a bit less
of a pain to specify SSA dependencies between cells.
* Also fixes various small bugs highlighted by this conversion.