[Arc] Add ZeroCount operation
Add an operation to represent leading and trailing zero counts. These
operations are commonly supported by modern processors and can allow
significant simplifications of the IR (see example below directly taken
from BOOM). The next step is a pass to convert these mux sequences to
the zero count and array operations.

```mlir
%2 = comb.extract %1 from 63 : (i64) -> i1
%3 = comb.extract %1 from 62 : (i64) -> i1
%4 = comb.extract %1 from 61 : (i64) -> i1
%5 = comb.extract %1 from 60 : (i64) -> i1
%6 = comb.extract %1 from 59 : (i64) -> i1
[...]
%64 = comb.extract %1 from 1 : (i64) -> i1
%65 = comb.xor %64, %true : i1
%66 = comb.concat %c-1_i5, %65 : i5, i1
%67 = comb.mux bin %63, %c-3_i6, %66 : i6
%68 = comb.mux bin %62, %c-4_i6, %67 : i6
%69 = comb.mux bin %61, %c-5_i6, %68 : i6
%70 = comb.mux bin %60, %c-6_i6, %69 : i6
%71 = comb.mux bin %59, %c-7_i6, %70 : i6
%72 = comb.mux bin %58, %c-8_i6, %71 : i6
[...]
%128 = comb.mux bin %2, %c0_i6, %127 : i6
```

This can now be converted to:

```mlir
%c1_i64 = hw.constant 1 : i64
%2 = comb.or %1, %c1_i64 : i64
%3 = arc.zero_count leading %2 : i64
%4 = comb.extract %3 from 0 : (i64) -> i6
%5 = hw.array_create [0, ..., 63] : i6
%6 = hw.array_get %5[%4] : !hw.array<64xi6>
```

Here the array can be entirely optimized away as it just encodes the
identity function to the index operand. There are other cases where the
element in the array is `1 << index` which can also be easily detected
in a second step.

These are 2-4 cycles runtime (LZCNT is 1 cycle on AMD, 3 cycles on
Intel) compared to the at least ~32 cycles of the mux chain and reduced
binary size.