[ESI] [mostly NFC] Lower ports pass refactoring (#4670)
Hopefully this refactor will 1) be more robust and rely less on implicit rules and 2) be more amenable to supporting new signaling standards. Makes a few understandability vs efficiency tradeoffs, but I think it's much easier to follow. It does add a good amount of code (and use more memory), but it's far simpler to reason about and separates the port assignment / instance-update-resolution from the signaling standard. It also has a ton of documentation since this is a relatively complicated pass.

This was mostly NFC, but it does change the order of output ports to something slightly more reasonable.