[Arc] Add ArcCanonicalizer pass
Add a pass that allows us to implement canonicalization patterns that
require a symbol cache to work efficiently (as regular canonicalization
patterns don't get one provided). But also allows us to add
canonicalizations to operations from other dialects that are simulation
specific and thus not desirable to be added to that other dialect
directly.