[FIRRTL] Fix firrtl-emit-metadata for SiTest BBox
Fix incorrect behavior for SiTest BlackBoxes when the
firrtl-emit-metadata pass is run without a MarkDUTAnnotation.  This
checks the existing working behavior (with a MarkDUTAnnotation, only
things exclusively instantiated by the TestHarness are put in the
TestHarness JSON file) and fixes/checks the non-working
behavior (without a MarkDUTAnnotation, everything is put under the
Design JSON).

This is not an exhaustive fix for all instances of this type of error.
This narrowly fixes the SiTest BlackBoxes case.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>