[FIRRTL] Fix Mem Metadata DUT/No-DUT Logic
Remove the "tb_seq_mems.json" metadata file as this has not been used
for a long time.  This was previously a metadata file that would list
memory information about the test harness.

Fix bugs in how memory metadata files are generated for situations where
no DUT is specified and when a DUT is specified, but a memory is
instantiated in both the DUT and the Testbench.

Previously, if no DUT was specified then all memory instances would be
placed in "tb_seq_mems.json".  (This is now gone.)  Now, all instances
will be placed in "seq_mems.json".

Previously, if a memory was instantiated in both the DUT and the
Testbench, then it would wind up with all instances placed in
"seq_mems.json", including the one from the Testbench.  Change this so
that only instances that are inside the DUT are included in
"seq_mems.json".

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>