[HW][NFC] Add missing InferTypeOpInterface for ArrayGetOp
The `ArrayGetOp` has a `TypesMatchWith` constraint which ensures that
the result type of the op is the element type of the operand. This
implicitly creates implementations for the `InferTypeOpInterface`, but
only if `mlir/Interfaces/InferTypeOpInterface.td` is included.
Explicitly add the include in `HWAggregates.td` and get rid of the now
obsolete builder. This ensures that if any other op in HW decides to
make use of `InferTypeOpInterface` we don't suddenly start seeing
unrelated failures due to `ArrayGetOp` suddenly picking this up.