[Support][NFC] Factor out ImplicitSSAName directive (#4774)
The FIRRTL and SV dialects have an almost interchangeable implementation
of the `ImplicitSSAName` directive, which they apply to the `attr-dict`
of the operation. To facilitate adding ops in other dialects that want
to use a similar scheme, factor the commonalities of these directives
out into the `CIRCTSupport` library.

The SystemC dialect uses a slightly different approach: it also uses a
`ImplicitSSAName` directive, but that one is defined in the HW dialect
and is applied to the `$name` of the op instead of the `attr-dict`. The
benefit is that TableGen automatically generates the list of attrs to be
elided (since `$name` is explicitly covered in the assembly format), and
printing the name in case it mismatches with the SSA name is handled
outside of the generic `attr-dict` printer. We might want to move our
ops to this approach in the future. The current refactoring tries to
keep changes to a minimum for now.

Co-authored-by: Hideto Ueno <uenoku.tokotoko@gmail.com>