[FIRRTL] make IMCP correct for registers and more aggressive (#4777)
Make IMCP not try to optimize registers. All the existing firrtl register optimizations are unsafe in that they change visible behavior. For this reason, they those optimizations are to be moved out to a separate pass and IMCP will be left as purely behavior preserving.

Secondly, many folders can operate on incomplete operand information. Rather than giving up and not folding these, if we have some constants, we try to fold, but we don't over-define if some of the operands were unknown. This way things like or(const 0, unknown) can fold.