[FIRRTL] Do not lower sub-access to -index on OOB
Change LowerTypes to not try to lower a sub-access to a sub-index
whenever it sees a constant.  Instead, first check that the constant is
in bounds and use the multi-bit mux lowering if it is not.

Previously, sub-index in-bounds index checking happened incorrectly
during result type inference.  This causes this illegal sub-index to
immediately error on construction (with a stack trace since type
inference failed).  This behavior wasn't great (#4753).  Now that
sub-index in-bounds index checking is handled via verification, this
verification is delayed until after LowerTypes runs.  This creates
problems for later LowerTypes user processing which wants to look
_through_ sub-index ops.  However, if the ops have illegal out-of-bounds
indices, this fails.  Fix all this by ensuring that only legal sub-index
ops are created during LowerTypes.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>