[ExportVerilog] Add hw.wire emission support (#4823)
Add support for the new `hw.wire` op to ExportVerilog. More specifically
this commit makes LegalizeNames consider `hw.wire` during name
legalization, and it adds a lowering from `hw.wire` to `sv.wire` to
the PrepareForEmission pre-pass. The rationale behind implementing HW
wire emission as a lowering in the pre-pass instead of direct support in
ExportVerilog is that it makes finding a location for the wire
declaration and its assignment an IR problem instead of something done
ad-hoc during emission, which makes it a lot easier to solve because we
can reason about order based on the ops instead of text snippets.