[IMCP] Make field sensitive (#4795)
This makes IMCP field sensitive. Previously lattice values were associated with `Value` but this PR changes to use FieldRef as Key. We could use `FieldSource::PathNode` as a key, but I choose FieldRef  mostly because of compile time performance reasons. With `FieldSource::PathNode`, we have to use a pair of `<PathNode::src, PathNode::path>`  as a key and it's necessary to iterate SmallVector to compute a hash of the key. Also in IMCP iterations, it's necessary to enumerate subelements of aggregates but we have to manipulate access paths to do that. Hence FieldRef was considered as a better choice because we can refer subelements by simple integer additions/subtractions. 

#### Current limitation
* Some operations which create aggregate values are not handled (vector_create or bundle_create). 
* Subaccess is not handled

#### Compile time change
There are several additional costs to use FieldRef as keys and to traverse each elements with `foreachFIRRTLGroundType` (we have to recurse based on types). 3% ~5% increases with the current implementation. 2% increase with a change of semantic of unknown states.
* BaseLine:
   `9.0791 (  1.2%)    9.0791 (  4.3%)    IMConstProp`
* New implementation (and w/o 1d vec preservation)
   `17.4909 (  2.1%)   17.4909 (  7.7%)    IMConstProp`
* New implementation (and w 1d vec preservation)
   `20.2466 (  2.6%)   20.2466 (  8.7%)    IMConstProp`