[Arc] Add clock_domain operation (#4917)
The arc.clock_domain operation is a dataflow equivalent of the
arc.clock_tree operation. It is useful to isolate clocks before lowering
state where the region still has to be a graph and arbitrary outputs
have to be supported. Several transformations will benefit from an
earlier isolation of clocks as they can often only be applied within a
single clock domain (e.g., cannot pack differently clocked arcs inside
the same vector for vectorization) and, as a result, don't have to
check the clock operand manually anymore.