[OM] Add support for primitive values. (#4840)
This adds a new operation for defining primitive values to use in
Classes according to the design in the rationale. Primitive values are
part of a simple expression grammar of what may be assigned to a
Field. Any typed Attribute can be used with the new operation to
create a primitive value of the Attribute's type. Some simple
round-trip tests are added, and some existing tests are simplified.