[CombFolds] Don't flatten operands when operand has a name hint.
This updates `tryFlatteningOperands` to check if an operand has a name
hint, and not flatten it if so. This weakens the optimizations in the
presence of name hints, which could lead to some missed opportunities
for constant propagation, etc. However, the optimizations are mostly
aesthetic at this level, and this might be a worthwhile tradeoff for
preserving names when requested.