[Arc] Memory ops: optional enables and folders (#4976)
Make the enables of the MemoryReadPortOp, MemoryWritePortOp, and MemoryWriteOp optional. Its absence means a constant true enable. Add folders to remove the operand when it is constant true as well as remove the entire operation when the enable is constant false as the writes do not ever happen in this case and the read returns an undefined value.