[HW] Add an enum comparison operation (#5032)
There is currently no way to compare the value of two enumerations as an
expression. There is an icmp operation, but it only accepts
IntegerTypes, and it would complicate things to allow enums to be used
everywhere an integer is allowed. The other way to compare enumerations
is using case statements, but it can't be used as an expression.

This PR adds an equality comparison operator for enumeration types.
Theoretically we should be able to support other types of comparisons,
such as <, since enumeration members are ordered. Since all I care about
is equality, this does not support those other kinds of comparisons.

It would also make sense to me to add a more generic equality operation
which could support comparison of structs, unions, and arrays as well as
enumerations.

I elected to implement exactly what I needed, since I feel like I can't
predict future needs here.