Revert Alternative Connect/Invalidate Parsing
This reverts the following commits:

  - 17e6c71fcd65c72ac51dd66f74b6a0a720832611
  - 2ffdfd86332c6d35813d32e93cc413a3d140ed8d

This has problems with situations where the leading token is either a
keyword or an identifier.  E.g., these two commits will cause errors on
ambiguous things like:

  connect is invalid

These will also cause errors on simpler cases like:

  connect <= foo
  invalidate <= bar

These need to do some checking to make sure that the token is not an
identifier before they proceed with assuming that it is the alternative
syntax.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>