[FIRRTL] Parse Connect/Invalid from Spec v3.0.0
Enable parsing of new Connect/Invalid syntax from the FIRRTL Specification
version 3.0.0.  This is _only_ enabled if the ingested FIRRTL indicates
that it is version 3.0.0 as the old syntax and new syntax are ambiguous if
mixed in the same file.

The old and new syntaxes for connect are:

    a <= b
    connect a, b

The old and new syntaxes for invalidation are:

    c is invalid
    invalidate c

Ambiguity arises when there exists an identifier called "connect" or
"invalidate".

This is a fixed-up revert of commit that removed simultaneous parsing of both
syntaxes:

  - c2251df4e5a86e2ed6f0cccd196bf380488e6baf.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>