[Arc][LowerState] Fix combinational loop false positives (#5068)
Fix the worklist that materializes values in clock trees to do a proper
depth-first traversal of the operands, instead of trying to push the
operands onto the worklist all at once. Doing the latter would lead to
issues where all operands of an operation would be pushed onto the
worklist, but then one of those operands might also depend on some of
those operands:

   %0 = comb.and %x, %y
   %1 = comb.or %0, %y
   arc.state @foo(%0, %1)

This would push `%0` and `%1` onto the worklist, but then when trying to
materialize `%1` the pass would falsely detect a combination loop when
it tries to materialize `%0` and sees that it's already in the worklist.

This fix now collects all nested operands of an operation and tracks
them alongside the operation, but then only processes the operands
one-by-one such that a proper depth-first pattern is traced out.