[FIRRTL] Add clock gate intrinsic (#5073)
Add a new `firrtl.int.clock_gate` operation to represent clock gates.
This new op is intended to eventually replace the magic `EICG_wrapper`
extmodule that some passes in the firtool pipeline are sensitive to
(mainly `WireDFT` and `ExtractInstances`).

This commit merely adds the op, canonicalization, basic round-tripping
tests, and the required snippets in `LowerIntrinsics`. It does not add
any handling of the intrinsic to the pipeline yet.

As a drive-by improvement it also makes the `UInt1Type` in
`FIRRTLTypes.td` a `BuildableType` and requiring an inferred width.
This allows `UInt1Type`s to be omitted from assembly formats. The `mux`
gets a shiny new `UInt1OrUnsizedType` to continue accepting uninferred
widths for the condition.