[FIRRTL] Isolate buildable types to known non-const cases
After discussion with @trilorez, revert the wholesale addition of
`BuildableType` to `ClockType` and friends, since this makes the types
buildable but defaulting to non-const. Instead the intention is to make
them *not* buildable and have the ops properly infer the types from the
constness of the operands. See discussion in #4742.

Create separate `NonConst*` flavors of a few types for use in ops where
they are known to be non-const, to reduce the type noise of those ops.