[OM] Add Reference attribute. (#5084)
This is a wrapper around a hw.innerNameRef with !om.ref type. This is
needed because the OM dialect constant ops need typed attributes, so a
hw.innerNameRef can't be used directly. That is really what we want to
capture, so this simply wraps the HW dialect attribute and provides a
type so it can work in the OM dialect type system.

Note that this does not encode hierarchical paths. This can be
accomplished in a variety of ways if necessary, but for now simply
referencing an inner symbol should be sufficient for many use
cases. If hierarchical paths are needed, the instance graph can be
elaborated and this information combined with the inner symbol
references.