[ExportVerilog] Automatically typedef all anon enums
Our Verilog output for anonymous enumeration is highly problematic:
every time we write out the type, the variants or members of the
enumeration enter the global namespace.  This makes it impossible to
have multiple declarations share the same enumeration type.

This adds walk of all operations in the ModuleOp to find and replace all
enumeration types used with typedefs.  The typedefs are automatically
injected into a typescope with the name "Enums".

Here is an example of what that looks like:
```verilog
typedef enum bit [0:0] {enum0_A, enum0_B} enum0;
module EnumCmp(
  input  enum0 test,
  output       result
)

  assign result = test == A;
endmodule
```

This also adds support for 0 width typedecls, which allow us to comment
out the zero width typedefs and the ports of that type.