[FIRRTL] Fix foo.0.0 parsing error
Change the FIRRTL lexer to make it context sensitive when parsing fields.
This is necessary for situations like "foo.0.0" where lexing after "foo."
will lex a float when it needs to lex an integer.  There are other ways of
fixing this on the FIRRTL specification side, but this is allowable by the
SFC today.

In Chisel, this enables usage of a MixedVec inside a MixedVec.

Fixes #5110.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>