[FIRRTL][LowerMatches] Fix matches with 0 and 1 variants (#5124)
This fixes two bugs in LowerMatches.

The first is when the match has no case statements, we would underflow
an integer and crash assuming too many regions.  This was fixed by just
deleting the match statement when there are no cases.

The second bug comes from a single case statement.  The code assumed
that we would be inserting in to the `else` block of a WhenOp, and that
it could replace the current block with the case block.  This is not
actually true in the case of a single case, where the current block and
insertion point is at the MatchOp.  We fix this by splicing the case
block in to the current insertion point.