[Arc] Add IsolateClocks pass (#5146)
Implement the clock isolation into the clock_domain op that got added earlier. Implement an interface that all clocked operations can inherit from such that passes can just work on that one and when we add a new clocked operation for example, we don't need to update the pass. Also we don't need to implement ugly helpers or template magic in every pass that wants do delete or get the clock of multiple clocked operations. In the future, we could also implement a ClockedRegion trait or interface to abstract the clocked op verifiers, etc.