[Arc] Generalize InlineArcs pass (#5144)
InlineArcs only considers arc.state operations at the moment. This adds
support to also inline arc.call sites. While the callgraph of the state
ops is entirely flat as they can only occur in a hw.module and there is
no hierarchy of such modules, the callgraph including arc.call
operations form a DAG. Arcs are not allowed to have recursions, but as
this is not properly verified, this pass checks for that and emits a
warning if such a case is deteced (and leaves the recursive structure as
is). In order to perform the least amount of calls to the inlining
utility, this pass is extended to traverse the callgraph bottom-up. This
way calls are inlined before they potentially get duplicated by inlining
in multiple call-sites.
Additionally, this pass removes all arcs that are not used now.
This also adds some options to the pass to control whether only states,
only calls, or both should be inlined as well as the maximum number of
non-primitive ops an arcs is allowed to have to still be eligible for
inlining.