[Arc] Simulation specific optimization for comb.icmp (#5198)
On the HW/Comb level, the `comb.concat` operation is basically considered free. However, in simulation it is quite expensive. Instead of concatenating integers of the same width, we can do the AND/OR directly and do a reduction on the smaller result integer, effectively cutting the number of assembly instructions produced to almost half.