[LowerSeqToSV] Fix a dominance bug when restoring subaccess (#5215)
There was a bug in LowerSeqToSV which creates invalid IR regarding dominance. Previously sv.array_index_out op might be created in a procedural region which is not a graph region. So this PR makes sure that sv.array_index_out is created in the toplevel so that we never create invalid IR.