[DC] Add merge operation (#5213)
`dc.merge` will select one of the incoming tokens and emit an output stating which token was selected. If multiple tokens are ready to transact at the same time, the tokens are selected with priority, from first to last (i.e. left to right in the IR). This property ensures deterministic behavior.