[FIRRTL][LowerXMR] Use ABI for refs when no internal path info. (#5229)
Ref-type ports on external modules should resolve to uses of the ABI macros when the path information isn't present (e.g., ref is "a.b.c" statements).

Note that the "circuit" name for an external module is expected (must) be the same as its module name. Without a mechanism to mark the circuit an extmodule appears under, we assume its in a circuit of that name.

These macros are used with expectation the tool/user will ensure the right ABI files are included.

cc #5218 , test is based on example there.