[OM] Add Evaluator API to get the underlying Module.
This API is useful when you have an Evaluator, but need to get back to
the core MLIR data structures. For example, to get a Location to print
an error, or to get the context for creating attributes. This change
will be useful in a couple upcoming patches, so committing it now as
"obvious".