[Pipeline] Switch latency insensitive mode to DC
Moves the latency-insensitive mode of `pipeline` up the abstraction ladder, from ESI to DC.
The motivation for this is that a latency-insensitive pipeline eventually will need to lower to something which contains control constructs (mainly join and fork operations). As such, by retargeting pipeline to DC, a lowered pipeline will eventually be able to mix-and-match with surrounding DC constructs (i.e. peephole optimization), before DC then lowers to ESI channels.