[Comb] Fix dominance issue in ICmp canonicalizer
The `comb.icmp` op has a pretty clever canonicalizer that refactors
`icmp.eq(xor(a, b, cst1), cst2)` into `icmp.eq(xor(a, b), csr1 ^ cst2)`.
If the xor on the LHS has other uses besides the icmp, a replacement op
`xor(xor(a, b), cst1)` is created which reuses the new `xor(a, b)` LHS
of the icmp.

These operations were inserted right before `icmp` instead of the `xor`,
which could have the unintended side effect of moving the `xor`s down
to just before the `icmp`. This is never an issue in `hw.module` because
it's a graph region. But in SSACFG regions this can lead to dominance
violations.

Add a test and set the rewriter's insertion poitn to the `xor`, such
that the replacement ops appear where they are supposed to be.