[ETC] Extract registers and instances if not in design (#5331)
This commit extends ETC to extract registers. 
In order to enable the same generic logic to extract registers, the
pass has been moved earlier in the firtool pipeline, so it can work at
the seq.firreg level of abstraction, which behaves just like an
instance. As firreg registers introduce cycles in dataflow by nature, this commit 
changes extraction mechanism so that we can find and extract maximum subset 
of testcode. Previously we were tracking dataflow until it reaches least fixedpoint 
of verification parts, but instead we can first compute least fixedpoint of design 
parts and subtract it from entire modules. 
---------

Co-authored-by: Mike Urbach <mikeurbach@gmail.com>