[FIRRTL] Initial pass to extract OM classes from FIRRTL modules.
Modules with input or output properties are converted to OM
classes. The input properties become parameters, and the output
properties become fields. The ops directly connected to output
properties are moved over. As needed, this can be extended to move
over whole dataflow slices, or adapted to work with ops besides
connects.

The property ports and operations are removed from the FIRRTL modules,
so after this pass runs, the modules only contain RTL.

This only handles the simplest cases: output properties driven from
input properties and constants. Handling instances of
property-containing modules will come in a future addition. This
includes fixing up instantiations of mutated modules to match their
declaration.

The OM type system is open, so this produces legal IR, but not IR that
is usable with the Evaluator. A future addition will include a pass to
convert to the operations, attributes, and types the OM dialect
expects to work with in the Evaluator.