[Arc] Split read latency into pre/post storage phases
Add support for more detail read latencies to the `InferMemories` pass.
It now supports a pre-storage-read and a post-storage-read latency,
allowing for more precise modeling when exactly the read synchronizes
with other writes. Memories only specify a single read latency. For now,
interpret that read latency as pre-storage-latency, meaning that the
address and enable signals are delayed and the underlying memory is only
read at the very end of the latency, such that all writes happening in
parallel are still visible. This can be changed easily in the future.