[FIRRTL] Support instances of property modules in ExtractClasses.
This finished support for handling InstanceOps with properties by
converting them into ObjectOps. The existing traversal that builds up
fields on ClassOps is maintained, but a special case is added for when
the property comes from an InstanceOp.

Using a cache to allow re-use, ObjectOps are created, and
ObjectFieldOps are created to extract their outputs. Some bookeeping
is necessary to ensure the relevant property ops feeding the instance
are cloned, the IRMapping is kept up to date, and the right property
ops are cleaned up. Otherwise, the rest of the pass is unchanged.