[Pipeline] Add `pipeline.latency` operation (#5340)
The `pipeline.latency` operation represents an operation for wrapping
multi-cycle operations. The operation declares a single block
wherein any operation may be placed within. The operation is not
`IsolatedFromAbove` meaning that the operation can reference values
defined outside of the operation (subject to the materialization
phase of the parent pipeline).

This commit includes changes to the register materialization pass
(differentiating between `regs` and `pass` inputs to stages) and
pipeline HW lowering. Currently, `pipeline.latency` operations are
just inlined into the current insertion point.