[FIRRTL][InferWidths] Handle foreign and prop types uniformly (#5392)
This changes InferWidths to use a common pattern to filter out foreign
and property types.  The `getBaseType(FIRRTLType)` function now takes a
`Type` and returns null if it is not a `FIRRTLType`.  `declVars` and all
three constrain functions now use `getBaseType` to skip over foreign and
property types.  This changes the behaviour of `maximumOfTypes` and
`unifyTypes`, but this is not a testable code path.