[FIRRTL][CheckCombLoops] Support property types (#5383)
This change adds support for property types to CheckCombLoops. Right now
property types show up as regular ports on instances and modules, and
are assigned the through the PropAssignOp, which implements the
ConnectLike interface. Module ports, instance ports, and connect like
operations had to be updated to support FIRRTLTypes instead of just
FIRRTLBaseTypes.