[HWModuleLike] Add by-name port lookup
Adds capabilities to the `HWModuleLike` interface to lookup in/output port indices by name. This is done by maintaining a mapping (`DictionaryAttr`) between port names and their indices.

The `arg/resIdxMap` attributes are not inherent attributes of module-like ops (not part of parsing). this makes sense to me to avoid duplicating information. Instead, the mappings are updated whenever:
1. a module-like op is parsed
2. a module-like op is built
3. a module-like op has its ports modified through `hw::modifyModulePorts` or any of the other functions which i managed to determine modifies ports.

If a user modifies the ports of a module outside the above locations, it is the responsability of the user to ensure that `updateModuleIndexMappings` is called.

The performance of this comes down to the performance of looking up a key in a `DictionaryAttr`.

A final detail is that these mappings will only be available if a module has unique in/output port names, given that it is currently legal to have duplicate names in the `argNames`/`resNames` attributes, as well as this being supported by exportVerilog.

Lastly, `FIRRTL` module-like ops do not implement this interface (for now) given the unified in/out port list.

As a first user, updates `PipelineToHW` to lookup clk/rst ports by name. I'm sure there's tons of other uses throughout the codebase.