[Seq] Add FirMem op (#5009)
Add a new `FirMemOp` to the Seq dialect to capture the semantics of
FIRRTL memories at the HW level. The intent behind this change is the
same as for `FirRegOp`, which captures the specifics of FIRRTL registers
but within the type system and abstraction level of the HW dialect.

Memories are represented as a declaration with the `seq.firmem` which
returns a value representing the memory itself. This value is of type
`!seq.firmem<12 x 42, mask 2>` which represents an opaque handle to the
memory; it encodes the depth (`12`) and width (`42`) of the memory, plus
an optional mask if the memory is to support masked writes.

Memory ports are represented as separate operations:

- `seq.firmem.read_port`
- `seq.firmem.write_port`
- `seq.firmem.read_write_port`

These operations take the memory itself as their first operand. The
remaining operands are the specific signals required for the
corresponding type of port: clocks, enables, masks, mode select, etc.
Read and read-write ports also return the read data as their result.
Enables and masks are optional to allow for more concise IR for simple
memory configurations.

Representing ports as separate ops in the IR has the advantage that
adding and removing ports is trivial; the standard CSE and
canonicalization even allow for collapsing duplicate ports and removing
permanently disabled ports. Separate ops also make building up memories
in the IR a lot simpler, since the operands and results associated with
a port are grouped together, instead of having to spread them across an
overall operand and result list on the overall memory. Getting a
memory's ports means walking the list of users of the declaration op,
which is very cheap since the only users are the port operations.

This commit merely adds the `FirMemOp` alongside some canonicalization
and general tests, but the op isn't used by any part of CIRCT yet. Use
in FIRRTL-to-HW lowering follows as a separate commit.