[FIRRTL] Quote literal ids in MLIR bundle fields
Change the MLIR printing/parsing of the MLIR representation of FIRRTL
bundle types to use quoted strings to represent numeric field names (which
must be represented as literal identifiers in specification FIRRTL).

Consider the following FIRRTL type:

    { `42` : UInt<1>, foo : UInt<1> }

Previously, the MLIR textual representation of this was:

    !firrtl.bundle<42: uint<1>, foo: uint<1>>

This commit changes that representation to

    !firrtl.bundle<"42": uint<1>, foo: uint<1>>

There are no changes to identifiers which have a non-literal
representation, e.g., "foo" above is unchanged.  The old MLIR format is
now rejected.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>