[FIRRTL][LowerXMR] Support ref.sub.  (#5551)
Implementation:

Extend the refSend list node to support recording indexing
along the resolution path, in addition to its current sole
use of tracking symbols of instances along the path.

When resolving, gather these indexing operations and generate
the appropriate string suffixes for each to narrow the origin
reference accordingly.

MemTap now use this for creating the right indexing suffix (under test).

Notes:

This is not the more general support needed once we support
input probes (and more complicated flow of probes through
the design) but build on existing requirements/assumptions
here and throughout compiler to get this working for the
more straightforward cases.

Known limitations:

Indexing through an input probe will cause resolution error.
  * Given input probes are rejected during parsing, and our avoidance
    of input probes in wiring, this should not be user-reachable.
  * Longer-term input probes can be rewritten to not exist,
    and "de-squiggling" the dataflow generally so that LowerXMR
    and other passes don't have to accomodate input probes.

Forcing part of an aggregate (continues to) cause entire aggregate
to be preserved.

Expects names/sizes of aggregates to be preserved through the end of the pipeline,
as they're hardcoded into the XMR strings + ABI contents.
This is fragile and should be replaced with placeholders so they can be substituted
after ExportVerilog.  Plan is to use per-field HW inner symbols for that portion,
once they're supported in HW and onward.