[LowerFirMem] Insert memory modules before first user (#5587)
Currently the `LowerFirMem` pass inserts `HWModuleGeneratedOp`s all the
way at the top of the file. This is problematic since passes like
FIRRTL's `LowerToHW` will produce an SV preamble with macro definitions
that the memories want to rely on. Inserting the memory modules at the
top of the file puts them before the macro definitions that they rely
on.

This commit changes the insertion point of memory modules, making them
appear as late as possible in the IR, basically just before the first
module that instantiates the memory. This ensures that the memory
modules are always placed after that preamble and makes their location
more logical within the IR.