[CI] Split wheel upload into stages (#5628)
This change splits the python wheel upload into 2 stages (separate jobs). In the
first stage, wheels are built and uploaded simply as artifacts through the
native GitHub actions upload-artifact action. This stage runs on the desired
matrix of configurations (manylinux, macos, etc.) and on all event triggers. In
the second stage, the corresponding wheels are downloaded (download-artifact)
and uploaded to PyPI. This stage only runs on ubuntu and only for tags or
nightly build (i.e. skipping workflow_dispatch).

The reason for this change is documented in #5420.