[HW] Add externalizeable interface and externalization pass
This is an attempt at adding a new interface to the HW dialect - the `ExternalizeableOpInterface`.
Using this interface, an operation is able to interact with a new pass, named `hw-externalize` to replace itself with instances of a user-provided external module.

The interface requires an op to describe:
1. its set of required ports
2. its set of optional ports
The above are communicated as a regular `hw::PortInfo`, with a type, a direction and a name. However, the name will be used as a key for a user-provided port mapping. This implies that:
1. it is the operation which dictates which ports it requires of its externalized module, by describing port identifiers and port types.
2. It is the user which decides the actual names of those ports by providing the mapping to the announced port identifiers.

As a result of this, clock gate externalization loses its default-portnames capability. If this is a big issue, let me know - personally i prefer to keep things explicit so i don't see a reason to hard-code in external module port names into a pass...

We should be able to fairly easily extend this model to module parameters using an identical method as for ports. Didn't do that for now, since i'd like to land this first. However, given module parameters, i'd expect this interface/pass to then also be applicable to ops which may exist with many different in/out types in the IR (i.e. the FIFO op, which one probably wants to externalize to a parameterized SV module).

As an example, I implemented the interface for the `seq.clock_gate` op:

```mlir
hw.module @Foo(%clock: i1, %enable: i1, %test_enable: i1) {
  %cg0 = seq.clock_gate %clock, %enable
  %cg1 = seq.clock_gate %clock, %enable, %test_enable
}

// circt-opt --hw-externalize="op=seq.clock_gate module-name=SuchClock instance-name=gated port-names=input=CI,output=CO,enable=EN,test-enable=TEN" %s

hw.module.extern @SuchClock(%CI: i1, %EN: i1, %TEN: i1) -> (CO: i1) attributes {verilogName = "SuchClock"}
hw.module @Foo(%clock: i1, %enable: i1, %test_enable: i1) {
  %false = hw.constant false
  %gated.CO = hw.instance "gated" @SuchClock(CI: %clock: i1, EN: %enable: i1, TEN: %false: i1) -> (CO: i1)
  %gated.CO_0 = hw.instance "gated" @SuchClock(CI: %clock: i1, EN: %enable: i1, TEN: %test_enable: i1) -> (CO: i1)
  hw.output
}
```

Now, the next question is whether this is actually the right approach. What i like about this approach is that it's very straight forward to bring any external SV module as long as it provides matching ports for what's published by the operation.
However, that might also be a big probem. As @teqdruid has pointed out in other discussions, externalizing an operation is a lowering, and as such, lowering logic shouldn't be embedded (re. dictated) by an operation.  The canonical example here would be that e.g. there's an operation A in CIRCT, you have your own module modA that implements the semantics of A, but it has a different interface than what op A publishes, meaning that there is a way to externalize A to modA given adequate shim logic is inserted - shim logic, which is dictated by modA and not A.
A counterargument to that could then be, that, you could implement said shim-logic manually in SV, such that it adheres to the published interface of A.