[ExportVerilog] Use wire for inlined op operand
Fix a bug where an assertion could fire in ExportVerilog due incomplete
wire creation in PrepareForEmission.  This bug occurss if an always inline
operation was duplicated, but its operand is not always inline.  This
creates a situation where, if you were to run PrepareForEmission _again_,
it would create a wire (and no assertion would fire).

Change the logic when recursing through operands of an always inline
operation to create this wire.

Fixes #5605.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>