[FIRRTL] Add ArrayRefAttr to FIRRTL
A number of ArrayAttr inherent attributes in the FIRRTL dialect have custom
accessors defined, because we want the getter to return an ArrayRef<Attribute>,
not an ArrayAttr.  This PR adds an untyped ArrayRefAttr that will generate the
accessors we want. Maybe someday this could be delivered upstream to MLIR, if
it proves useful to us.