[HW] Consistent HWMemSimImpl ignore-read-enable
Fix an inconsistency between the handling of read ports and read-write
ports when running HWMemSimImpl with the "ignore-read-enable"
option (firtool option "-ignore-read-enable-mem").  After #5700, this
would cause the enable line of a read-write read port to have no effect.
I.e., the memory would always read even if not enabled.  Change this to
align with the behavior of read ports where the memory always returns the
last read address.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>