[SVExtractTestCode] Fix a bug that instance inlining doesn't update inner symbols and clone sv.bind (#5679)
Close #5665. This commit fixes a bug that
bound instances are accidentally erased. When inlining input only modules it's necessary
to clone bind statements for bound instances in the module. Previously single bind op is
shared by multiple instances, as a result ExportVerilog only emits a bind statement for one of them.