[OM] Implement OM linker pass (#5743)
This implements LinkModulesPass that is a core logic of `om-linker`.
The input IR consists of multiple mlir::ModuleOp that have their own
symbol namespaces. LinkModulesPass first constructs a map from
symbols to ClassLike operations and checks whether it’s legal to link the classes.

We consider a symbol used as an external module to be "public" thus we
cannot rename such symbols when symbols collide. We require a public symbol
to have exactly one definition so otherwise raise an error.

Also a definition and declarations must have the same input types and fields.