[Support] Add KeywordBool custom parser/printer directive
Add the `custom<KeywordBool>` parser/printer directive that allows for
booleans to be parsed and printed based on two keywords, one being
interpreted as "true", the other as "false".

For example,
```
custom<KeywordBool>($async, "\"async\"", "\"sync\"")
```
sets the property `$async` to true if the keyword `async` is found, or
false if the keyword `sync` is found. Otherwise an error is reported.