[firtool] Load Verif/LTL dialects, lower Verif to SV; NFC
Accept operations from the Verif and LTL dialects in firtool, and call
the Verif-to-SV lowering as part of the pipeline. This will allow
firtool to properly handle the new message formatting and verification
operations even when they are not generated by the FIRRTL part of the
pipeline.

Doesn't affect existing flows, but makes firtool not crash when used on
MLIR inputs that contain verification ops.