[ci] Skip tests for scheduled firtool publishing
Skip tests when creating nightly firtool binaries if using the "scheduled"
event.  This is already getting tested with the nightly integration tests in a
much more robust way and it's pointless to duplicate the tests here.

This gets the number of build steps down to ~1400 and causes binaries to upload
in 25 minutes.

Note: getting ccache setup on this job (and across all jobs) would help with
likely avoid the need for skipping tests.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>