[ci] Add Reusable Build/Test/Install Action, WIP
Add a work-in-progress reusable GitHub Action to do a unified build, test,
and install of CIRCT.  This is intended to be a replacement for pieces of
other GitHub actions which are doing similar, but slightly different
build/test/install.

This is directly landed onto main so that I can test this with a
workflow_dispatch event (manual trigger).  This even can only be used for
workflows that are on the default branch.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>