[FIRRTL] Add HierPathCache to reuse HierPathOps (#5909)
This change takes a common pattern in LowerAnnotations, which avoids creating
duplicate HierPathOps by maintaining a cache, and moves it into a dedicated
class for easier re-use.  Since the HierPathCache maintains an insertion point
instead of always inserting to the top of the circuit's block, HierPathOps are
now in reverse order after LowerAnnotations.