[FIRRTL] getOrAddInnerSym: fix bug when targeting a field (#5978)
When an `InnerSymbolTarget` targets a field, `target.isOpOnly()` returns false.
`getOrAddInnerSym` was mistakenly using this function to determine if the
target was a port or a regular operation, causing it to treat all targets with
a fieldID as port targets.  This would crash when casting the target operation
to a module op.  This changes `getOrAddInnerSym` to use `!target.isPort()`
instead of `target.isOpOnly()` to check if the target is a port.