[SV][HW] Stable port order in SVExtractTestCode
Change the BFS of operations used to compute backwards slices in
SVExtractTestCode to, instead, use a DFS of operands.  This ensures that
any intermediary operations which may show up do not cause the ports of
extracted modules to change.

This was triggered by addition of the `seq.from_clock` operation to cause
the BFS discovery order of roots to change.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>