[Pipeline] Remove outlined lowering (#6111)
Will be removing the outlined pipeline lowering. It's dumb to have to support two separate, but similar lowerings. Instead, rely on the inline lowering, and down the road, implement a "stage extraction" pass to extract operations that reside within the same logical stage after-the-fact by annotating ops during inline conversion.