[Ibis] Add ibis reblock pass  (#6089)
Recreates `ibis.block` operations from a CFG. Any `ibis.block.attributes` operations at the parent operation will be added to the resulting blocks.

The IR is expected to be in maximal SSA form prior to this pass, given that the pass will only inspect the terminator operation of a block for any values that are generated within the block. Maximum SSA form thus ensures that any value defined within the block is never used outside of the block.

It is expected that `ibis.call` operations have been isolated into their own basic blocks before this pass is run. This implies that all operations within a block (except for the terminator operation) can be statically scheduled with each other.

Also adds changes to the Ibis method op to make it more "imperative". This is under the assumption that we want to use the method op as the container of choice while going through "impeartive" lowering steps (i.e. all steps wherein we expect to keep around `cf` and MLIR blocks).