[FIRRTL] Add "removable" methods to Annotations
Add member functions to Annotation and AnnotationSet that can be used to
query if a specific Annotation or if all Annotations can be
removed/deleted.  There are certain "weak" annotations used by Object
Model which can be safely deleted because their downstream consumers will
_NOT_ error if they no longer exist.

As an aside, the whole notion of a removable annotation is somewhat
tenuous as a compiler could legally drop a removable annotation.  The
actual semantics of a removable annotation are best effort.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>