[FIRRTL] Update LowerClasses to use the defname for ExtModules. (#6243)
While lowering an ExtModule, if it has a defname set, use that for the
lowered ExtClass name. This is needed to ensure the lowered ExtClass
lines up with the lowered Class.

This also plumbs around the SymbolTable, so we can use it to
efficiently look up the referenced module while lowering instances,
and check that for a defname when we build the name of the class for
the object instance.