[Arc] Add LowerArcsToFuncs pass (#6227)
Extract the Arc-to-Func lowering from the LLVM lowering pass and make it
a separate step in the pipeline. This gives us a clear part of the
Arcilator pipeline where arcs have been lowered into functions, and the
entire simulation model is represented as `func.func` ops and the
corresponding calls.