[OM] Add more path types (#6250)
This adds more path types which will be used for greater safety in OMIR. This
change splits regular paths in to BasePathType and PathType, where the
BasePathType is only capable of representing paths through the instance
hierarchy, without being able to target hardware components such as wires and
registers. In the future, all paths can only be created relative to some base
path. This also adds a FrozenBasePathType and FrozenPathType, which will be
used to represent lowered paths which no longer depend on the hardware hiearchy
being present in the IR.