[FIRRTL] Update ExtractInstances to use the new instance symbol. (#6327)
In the case we need need to make a new NLA for an instance that has
moved past the module at which the old NLA was rooted, use the new
instance's inner symbol name. This code path used the original inner
symbol in the old NLA, but this would get out of sync if the new
instance needed to get a new symbol in the namespace. Similar logic
exists later in this method, but seems to have been missed here.