[Arc] Align >=128-bit integers to 16 bytes (#6366)
This is a fix for an issue introduced by llvm-project:a21abc7 via
d83d1a4. The alignment in LLVM was changed from 8 to 16 bytes but in Arc
we still allocated memory according to the old alignment, thus leading
to overlapping memory accesses (a store to a HW memory could modify the
state of a register stored after it in memory).