[Arc] Fix InlineArcs pass performance (#6379)
In a previous change we generalized the inliner to also be able to inline arcs into other arcs (see arc.call operation). There we also migrated our custom implementation to use the MLIR InliningUtils instead. However, it turned out that it performs poorly because of a lot of block splicing.

Note that the old pass traversed the call-graph bottom-up (inline arcs that don't have other calls inside first), while the current implementation processes arcs in the order in which they appear in the IR. However, the way it is implemented makes it easy to pre-compute any other ordering and pass it to the inliner.

Also, I slightly changed the way single-use and trivial arcs are counted in the statistics (now it just collects the statistics on the initial state of the IR instead of taking into account the inlining decisions of previously processed Arcs). The reason why there are that many more single-use arcs is that previously single-use arcs that were also trivial were only counted as trivial, but now they are additionally counted as single-use as well.