Bump LLVM (#6404)
The upstream change to !llvm.ptr has triggered a lot of churn in the LLHD side of CIRCT. We currently don't rely on llhd-sim for anything in-tree. A few breakages remain in the interaction between the llhd-sim runtime and the generated code:

* test/Dialect/LLHD/Simulator/sim_formats.mlir
* test/Dialect/LLHD/Simulator/sim_process.mlir
* test/Dialect/LLHD/Simulator/sim_reg.mlir
* test/Dialect/LLHD/Simulator/sim_wait.mlir

These are now gated behind a REQUIRES: llhd-sim-fixed in order to unblock the LLVM bump and allow for them to be fixed in parallel. We might want to introduce a few intermediate ops in the LLHD dialect to reduce the pretty big abstraction gap between LLHD ops and LLVM. A lot of things can now go through the SCF dialect (and maybe memrefs?).