[FIRRTL] Enable lowersigs and passive wires
This is a breaking change to chisel.  Chisel outputs don't touch on aggregate ports, which this rejects if the port requires lowering.  Don't touch and type lowering are incompatible.  The implementation of dontTouch in Chisel has been changed to default to marking leaves of aggregates instead of the root.  This will result in the same behavior currently observed while enabling future emission changes.