[Scheduling] Define problem to model operator chaining in cyclic problem. (#6485)
This PR defines a ChainingCyclicProblem, which models a hybrid problem of ChainingProblem and CyclicProblem and adapts the simplex scheduler to solve this problem. It is mainly done by reusing and adapting codes from the two base problems. This problem represents the problem a static HLS tool will solve for loop pipelining.

---------

Co-authored-by: leothaud <dylan.leothaud@irisa.fr>