[OM] Add integer shift right op.
This op shifts its lhs OMIntegerType operand right by the rhs
OMIntegerType operand to produce an OMIntegerType result. This defines
the evaluateIntegerOperation interface method in terms of APSInt >>.

If the shift amount is negative, an error is raised according to the
definition of the operation semantics.

Since the implementation simply uses the getExtValue of the shift
amount, this also raises an error if the shift amount does not fit
into an int64_t.