[NFC] Remove test demonstrating illegal IR.
This test appears to be testing ExportVerilog for a pattern that isn't
actually legal, since a `hw.type_scope` defines a Symbol. These days,
the relevant ifndef guard is emitted as a verbatim, for the exact
reason that the `hw.type_scope` must be in the top-level module's
SymbolTable.

Fixes https://github.com/llvm/circt/issues/6669.