[FIRRTL] Use untyped propassign source accessor in LowerClasses. (#6690)
When lowering object instantiations, we need to get the source from a
propassign to pass that Value to the newly created OM dialect object
op. However, in some cases, like when passing references to objects
down the hierarchy, we have already converted the source Value to an
OM dialect object op. This means the typed getSrc accessor will fail,
as it expects the source Value to be a FIRRTLType.

The IR is in the desired state at this point, so instead we can simply
use the untyped getSrcMutable accessor to get the corresponding
OpOperand, and get the Value from there. This avoids errors when the
pass has knowingly put the IR into a state that won't pass the
verifier.