[FIRRTL] Walk nested regions in SFCCompat
Fix a bug in the SFCCompat pass where only top-level operations in a
module were walked.  This resulted in SFCCompat behavior not running on
any nested regions.  This had no practical effect as the only regions that
have historically existed are in WhenOps and SFCCompat is always run after
ExpandWhens.  However, LayerBlockOps _do_ exist at this point in the
pipeline and SFCCompat should handle these.

This helps with #6740, but does not address it specifically.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>