[FIRRTL] Put layer collateral in testbench dir
Change the LowerLayers pass to put generated modules inside the testbench
directory if one is specified and a DUT is present.  This is a stopgap
measure that keeps layers aligned with existing CIRCT compilation of
FIRRTL.  This is intended to be replaced with a better FIRRTL ABI based on
filelists in the future.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>