[Moore] Rename files, add dialect rationale, fix docs
Rename `MIROps.{cpp,h}` to `MooreOps.{cpp,h}`. This makes the dialect be
more inline with others. We can always create additional dialects or
groups of operations later.

Also add the missing `moore` dialect rationale doc and fix the generated
documentation. Non-functional change in terms of operations and types,
but the documentation changes.

This is a preparatory cleanup step for the Slang upstreaming work.