[HGLDD] Fix instance output port emission
The `EmitHGLDD` translation currently emits instance output ports as XMR
`<instName>.<portName>`. This is not valid HGLDD. Remove this special
handling of instances. Instead if a value cannot be emitted as HGLDD,
try to follow the value's use as input to a wire or assignment and use
the destination name instead.

Also fix an issue where ops with a single result and `hw.verilogName`
attribute would be emitted as signal references with that Verilog name.
This code path would trigger for instances with a single output port,
causing the instance name to be used for the signal reference. Limit
that code path to only trigger on Comb dialect ops.

Fixes #6735.