[LowerToHW] Precompute modules under DUT
Change LowerToHW to precompute which modules are under the DUT during the
serial execution portion as opposed to dynamically querying the
InstanceGraph during the parallel execution portion.  This enables
querying the underlying properties of an instance to determine if it is
considered "under the DUT" which is not possible if done during the
parallel execution portion when instances are being erased.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>